refactor: use weekly entrance + nearest waypoint for all revive paths
Move _revive_goto_weekly_entrance from DomainTask to BaseCombatTask.
Both TacetTask (open world) and DomainTask (dungeon) now follow:
  ESC(close popup) -> F2 weekly entrance -> teleport_to_heal

DomainTask adds domain exit step before the shared weekly route.
This is more stable than M(map) directly after death state.
